Adding Mention Controller
