+++ server/app.js
@@ -3,7 +3,6 @@ const express = require("express");
 const { join } = require("path");
 const cookieParser = require("cookie-parser");
 const logger = require("morgan");
-const mongoose = require("mongoose");
 const indexRouter = require("./routes/index");
 const mentionRouter = require("./routes/mention");
 const pingRouter = require("./routes/ping");
